requirements.txt
python-telegram-bot[all]==21.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
apscheduler==3.10.4
pytz==2024.1
requests==2.31.0
gunicorn==21.2.0